+++ crates/adapters/src/format/json/mod.rs
@@ -43,6 +43,23 @@ pub enum JsonUpdateFormat {
     #[serde(rename = "debezium")]
     Debezium,
 
+    /// Format used to output JSON data to Snowflake.
+    ///
+    /// Uses flat structure so that fields can get parsed directly into SQL
+    /// columns.  Defines three metadata fields:
+    ///
+    /// * `__action` - "insert" or "delete"
+    /// * `__stream_id` - unique 64-bit ID of the output stream (records within
+    ///   a stream are totally ordered)
+    /// * `__seq_number` - monotonically increasing sequence number relative to
+    ///   the start of the stream.
+    ///
+    /// ```json
+    /// {"PART":1,"VENDOR":2,"EFFECTIVE_SINCE":"2019-05-21","PRICE":"10000","__action":"insert","__stream_id":4523666124030717756,"__seq_number":1}
+    /// ```
+    #[serde(rename = "snowflake")]
+    Snowflake,
+
     /// Raw input format.
     ///
     /// This format is suitable for insert-only streams (no deletions).
@@ -105,7 +122,7 @@ pub struct DebeziumPayload<T> {
 }
 
 /// A data change event in the insert/delete format.
-#[derive(Debug, Deserialize, Serialize, PartialEq, Eq)]
+#[derive(Debug, Deserialize, Serialize, PartialEq, Eq, PartialOrd, Ord)]
 #[serde(deny_unknown_fields)]
 pub struct InsDelUpdate<T> {
     // This field is currently ignored.  We will add support for it in the future.
@@ -121,6 +138,24 @@ pub struct InsDelUpdate<T> {
     /// from the table.
     #[serde(skip_serializing_if = "Option::is_none")]
     delete: Option<T>,
+}
+
+#[derive(Debug, Deserialize, Serialize, PartialEq, Eq, PartialOrd, Ord)]
+pub enum SnowflakeAction {
+    #[serde(rename = "insert")]
+    Insert,
+    #[serde(rename = "delete")]
+    Delete,
+}
+
+#[derive(Debug, Deserialize, Serialize, PartialEq, Eq, PartialOrd, Ord)]
+#[serde(deny_unknown_fields)]
+pub struct SnowflakeUpdate<T> {
+    __stream_id: u64,
+    __seq_number: u64,
+    __action: SnowflakeAction,
+    #[serde(flatten)]
+    value: T,
 }
 
 // TODO: implement support for parsing this format.