+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   errors during lifecycle state changes.
   ([#869](https://github.com/feldera/feldera/pull/869))
 
+### Added
+
+- Experimental Snowflake sink
+  ([#774](https://github.com/feldera/feldera/issues/774))
+
 ## [0.1.7] - 2023-10-10
 
 ### Added